+++ src/test/java/org/bigml/binding/RunCukesTest.java
@@ -33,6 +33,7 @@ import cucumber.junit.Cucumber;
         "src/test/resources/test_25_correlation.feature",
         "src/test/resources/test_26_statistical_test.feature",
         "src/test/resources/test_27_logistic_regression.feature",
+        "src/test/resources/test_28_associations.feature",
         "src/test/resources/test_29_script.feature",
         "src/test/resources/test_30_execution.feature",
         "src/test/resources/test_31_library.feature",